// frontend/src/types/clips.ts
interface HasId {
  _id: string;
}

export interface MetaMeta {
  created?: string;
  createdBy?: string;
  updated?: string;
  updatedBy?: string;
}
// ...
interface ClipConstruct extends HasId, MetaMeta {
  clip: string;
  label: string;
  description: string;
  timecodes: Array<[number, number]>;
}

export interface MusicBasic extends ClipConstruct {
  type: 'Music';
}

export interface Music extends MusicBasic {
  annotations: Record<MusicRelation, string[]>;
}

export interface SpeechBasic extends ClipConstruct {
  type: 'Speech';
}

export interface Speech extends SpeechBasic {
  annotations: Record<SpeechRelation, string[]>;
}

export interface NoiseBasic extends ClipConstruct {
  type: 'Noise';
}

export interface Noise extends NoiseBasic {
  annotations: Record<NoiseRelation, string[]>;
}

export interface PictureBasic extends ClipConstruct {
  type: 'Picture';
}

export interface Picture extends PictureBasic {
  annotations: Record<PictureRelation, string[]>;
}

// interface SegmentContains {
//   [elementId: string]: { [relation: string]: SegmentMember[] };
// }

export interface SegmentMember {
  added: string;
  addedBy: string;
  annotation: string; // annotation id
}

export type ElementType = 'Segment' | 'Music' | 'Speech' | 'Noise' | 'Picture';

export interface ClipElement extends ClipConstruct {
  type: ElementType;
  clip: string;
}

export interface Segment extends ClipElement {
  type: 'Segment';
  timecodes: Array<[number, number]>;
  segmentContains: SegmentMember[];
}